A bit nicer
You have to check the way to upload to make the download dependent on that. i.e. download a binary file and upload it without saving in the hdd in between. Currently, if you try to get a binary file it's going to fail horribly (cannot be parsed), maybe a new bget method?